Modified chr() --> bytearray() in code
In Python3, the concepts of text and (binary) data are handled differently
than in Python2 -- see Python docs.

The (mutable) 'bytearray' type is convenient to store data to be sent to
(and received from) the board. In addition, the type change will require
less checks, as bytearray is always a series of bytes: improper commands do
not raise error but are not processed (and will be dropped by the board,
probably).