Correctly process incoming bytes by board.iterate()
When testing the py3 code on Arduino Duemilanove, handling of
incoming board messages did not work: 'data' variable should
be of integer and not of byte type.

Tests are OK on Arduino: analog, servo, PWM, digital output.
All are physically working as expected with this commit.

However, the TestSuite gets broken with this commit.
The tests' read/write functionality must be checked for
read/write bytes only (as Firmata does in the real world):
as of now, I noticed also int/list/str type messages send
to the Testsuite queue.