Change the internal data structures to store Pin and Port data
In this commit, a dict is introduced to `board` instances,
which keeps track of Pin instances rather than managing
a list of Pin numbers. Also, Port instances are now stored
in a dict.

There is no change in the pyfirmata API: all commands will
work as before, without modified input/output.

The one, minor, exception on this is when a Pin instance is changed
to SERVO mode. Formerly, the mode change was followed by a
board.servo_config() command, which send the servo to the
default angle of 0. As this is unwanted behavior in some
use-cases, I have removed this automatic 'servo_config'
routine from Pin._set_mode = SERVO. The servo_config() function
must now be called explicitly, which gives the user control
on what initial position the servo must be set to.

This commit makes possible:
- implementation of autodetect, autoquery, etc. functionality
- that analog pins can be used as digital pins. Now,
  board.analog_pins[0] and board.pins[14] both point to the
  same Pin instance on an Arduino
- status information is now stored by the Pin and Port
  instances, and not by the Board.
- This commit should make it much easier to define Ports for
  non-optimized boards such as the Arduino Mega, as now, a
  Port instance does not need to have a sequence of Pins:
  a Port can be defined as having pins [1,2,3,7,8,9,10,11]
  for example.

The core change of this commit is 3 lines in `pyfirmata.board`,
where three dicts are added (self.pins, self.ports, and self.analog_pins).
The modifications in the rest of the code are due to a
rewrite of how Pin and Port information is accessed.

Apart from this core change, the pyfirmata.board.exit()
function has been updated to clear the pin/port information.
This rather small change enables tests to run well (previous
tests were thought to be with Arduino but in reality were ran
with a MockupBoard Arduino Mega for some tests -- as data got
appended to a board (more pins). This should also solve the
'FIXME' in the tests.py file which mentions that previous
data is stored in a 'fresh' board instance.