Correct set membership testing: does not work in Python2.6
A new test was added to the `board.get_pin()` function,
whichchecks if the first letter of the string is either an 'a'
(for analog) or a 'd' (for digital). If there's a typo here,
this check would catch it.

However, the test `if a_d not in {'a', 'd'}` does not work
under Python2.6. This (and two more `set` membership tests)
were replaced by a list membership test to satisfy Py2.6