Correct the test for 'a' or 'd' in the get_pin() function
A new test was added to the `board.get_pin()` function, which
checks if the first letter of the string is either an 'a'
(for analog) or a 'd' (for digital). If there's a typo here,
this check would catch it.

However, the test `if a_d not in {'a', 'd'}` does not work
under Python2.6. By changing the set to a list, hopefully
this now works under Py2.6.