Implementation pinList2boardDict Implementation of pinList2boardDict pyfirmata/boards.py -> 'use_ports' : True - Removed because its was used nowhere.
Making the tests pass:
I had to add every [127] pin as a digital pin

Houston, we have an API problem:
- (14, 15, 16, 17, 18, 19)
+ (0, 1, 2, 3, 4, 5)
Since the introduction of servo and i2c pins, we can not
state that i2c is always a digital or analog pin. On
Arduino they are analog(A5, A4), and on Mega they are digital(20, 21).