Fix the exception raised when a timeout happens in iterate().
Previously some timeouts (when reading from the board and not enought
data was available) would raise a
"TypeError: ord() expected a character, but string of length 0 found"
exception, that is not very useful to know what's happening.

Now it will raise a
"IOError: Failed to read data from port, read N bytes expected M bytes"
that is easier to understand and shows more information.