Add 'with' statement support to the Board class.
This will avoid problems with the exit() method not being called
(the __del__ exit() statement was not always called! causing the
connection with the board to hang if an Iterator thread was running).

Also it simplifies the code, as an Iterator will be managed
automagically for the with block.

board = Arduino('/dev/...')
with board: # an Iterator gets started automagically.
    board.get_pin('d:8:i').read() # Works! as an Iterator is running.
    # board.exit() gets called automagically and the Iterator stops
    # after this line.