Flag Iterator threads as 'daemon' to prevent hangs at shutdown.
"The significance of this flag is that the entire Python
program exits when only daemon threads are left."

This way Python won't hang at exit, but will just warn of
an exception at shutdown instead.

Anyway it's still better to call ``Board.exit()`` or use
a ``with board: ...`` block to guarantee proper shutdown
and avoid this warning.